### 2021-02-20 #3047 @whedon
[PRE REVIEW]: WhaleMap: a tool to collate and display whale survey results in near real-time
**Submitting author:** @hansenjohnson (<a href="http://orcid.org/0000-0002-3086-6759">Hansen Johnson</a>)
**Repository:** <a href="https://github.com/hansenjohnson/WhaleMap" target ="_blank">https://github.com/hansenjohnson/WhaleMap</a>
**Version:** v1.0.0
