### 2021-02-20 #3047 @Kevin-Mattheus-Moerman
@hansenjohnson thanks for this submission. Our system detected the paper currently lacks a `Statement of need` section [which is required](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Can you add such a section to you paper? It should illustrate the research purpose of the software.
