### 2021-02-20 #3047 @Kevin-Mattheus-Moerman
Notes for handling editor, this work seems to relate to these papers:
https://github.com/openjournals/joss-reviews/issues/1825
https://github.com/openjournals/joss-reviews/issues/2097
https://github.com/openjournals/joss-reviews/issues/2172
