### 2021-02-25 #3047 @kthyng
Hi @hansenjohnson and thanks for your submission. It looks like a really interesting project! I am going to ping the editorial board to make sure that it is in scope before we move forward. This isn't a statement on how awesome your project is, but just to check for fit. This will take 1-2 weeks and someone will get back to you. Thanks for your patience.
