### 2021-03-12 822da6e
specify file start to reject open excel files
