### 2021-03-15 a4b494a
use IDREL to define score
