### 2021-03-15 1b447f3
drop sightings with 0 as number (used to mark dups)
