+++ R/proc_tss.R
@@ -1,4 +1,15 @@
-wrangle_tss = function(file_name, lon_col, lat_col, lstart, lend, pstart, pend, plot_tss = F, fig_dir = 'figures/tss/'){
+## proc_tss ##
+# process traffic separation scheme data
+
+# setup -------------------------------------------------------------------
+
+# libraries
+suppressPackageStartupMessages(library(rgdal))
+suppressPackageStartupMessages(library(tools))
+suppressPackageStartupMessages(library(raster))
+
+# canadian tss reader
+wrangle_tss = function(file_name, lon_col, lat_col, lstart, lend, pstart, pend){
   # This function will, with some supervision, parse a 'tss file' to extract the lat-lon positions of lines and polygons in a consistent format. It will also optionally attempt to plot the processed tss coordinates.
   # file_name -> path to input file ('.csv')
   # lon_col -> column index (integer) of csv where longitude values are stored
@@ -6,13 +17,7 @@ wrangle_tss = function(file_name, lon_col, lat_col, lstart, lend, pstart, pend, 
   # l/pstart, l/pend -> start and stop row indices of polygons (p) or lines (l)
   # plot_tss -> switch to plot
   # fig_dir -> directory to save figures
-  
-  # libraries
-  library(tools, quietly = T, warn.conflicts = F)
-  library(oce, quietly = T, warn.conflicts = F)
-  library(ocedata, quietly = T, warn.conflicts = F)
-  data("coastlineWorldFine")
-  
+
   # read in data
   f = read.csv(file = file_name, header = F, colClasses = 'character')
   
@@ -52,29 +57,13 @@ wrangle_tss = function(file_name, lon_col, lat_col, lstart, lend, pstart, pend, 
     ld = NULL
   }
   
-  # plot tss
-  if(plot_tss){
-    if(!dir.exists(fig_dir)) dir.create(fig_dir, recursive = T)
-    
-    png(paste0(fig_dir, name, '.png'), height = 5, width = 5, units = 'in', res = 100)
-    plot(coastlineWorldFine, 
-         clon = clon,
-         clat = clat,
-         span = span
-    )
-    lines(ld$lon, ld$lat)
-    polygon(pd$lon, pd$lat, col = 'blue')
-    mtext(side = 3, adj = 0, text = paste0(basename(file_name)))
-    dev.off()
-  }
-  
   # combine into data frame list
   out = list(tss_lines = ld, tss_polygons = pd)
   
   return(out)
 }
 
-# process each tss files --------------------------------------------------
+# process canadian TSS ----------------------------------------------------
 
 # initialize list for gathering output
 TSS = list()
@@ -95,10 +84,6 @@ TSS[[3]] = wrangle_tss(ifile, lon_col = 5, lat_col = 6, lstart = 65, lend = 112,
 ifile = 'data/raw/tss/Boston.csv'
 TSS[[4]] = wrangle_tss(ifile, lon_col = 1, lat_col = 2, lstart = 6, lend = 16, pstart = 0, pend = 0)
 
-# # canso - duplicated with chedabucto
-# ifile = 'data/raw/tss/Canso.csv'
-# TSS[[5]] = wrangle_tss(ifile, lon_col = 1, lat_col = 2, lstart = 20, lend = 41, pstart = 5, pend = 16)
-
 # chedabucto
 ifile = 'data/raw/tss/Chedabucto_bay.csv'
 TSS[[6]] = wrangle_tss(ifile, lon_col = 1, lat_col = 2, lstart = 2, lend = 24, pstart = 27, pend = 48)
[…]
