### 2021-03-25 cf35154
change from DMA to slow zone
