### 2021-03-26 674c33f
add rgeos dependency to proc_gis
