### 2021-04-02 #3094 @pjbouchet
WhaleMap is cited/mentioned in more papers than currently reported in the reference list, though some authors only provide the app URL rather than include an actual citation which could be more easily tracked. You may want to add the following to your list of references:

- [Koubrak et al. 2021](https://www.sciencedirect.com/science/article/pii/S0964569120300193). Saving the North Atlantic right whale in a changing ocean: Gauging scientific and law and policy responses.

- [Gervaise et al. 2021.](https://www.sciencedirect.com/science/article/pii/S0003682X21000426) Optimizing passive acoustic systems for marine mammal detection and localization: Application to real-time monitoring north Atlantic right whales in Gulf of St. Lawrence.

- [Baumgartner et al. 2020. ](https://www.frontiersin.org/articles/10.3389/fmars.2020.00100/full) Slocum Gliders Provide Accurate Near Real-Time Estimates of Baleen Whale Presence From Human-Reviewed Passive Acoustic Detection Information.
