### 2021-04-09 #3094 @pjbouchet
I don't have any specific comments, other than the above. The app is well designed and works well when I tested it online. A couple of thoughts and suggestions for future improvements:
- I wonder about the sustainability of the workflow given the need to have custom scripts written by the curator for each survey team. In my experience, "allowing any team to submit data in nearly any format" (as stated in the article) is a recipe for headaches. What happens if the curator is suddenly unavailable, changes jobs etc.? Are there systems in place to ensure that the running of the app remains unaffected? I would argue that encouraging data standardisation during collection would be worthwhile.
- Are you thinking / have you thought about adding environmental layers to the map (and the possibility to query environmental data at the locations where whales were detected)? This would be interesting and could support spatial modelling efforts to better understand how whale distribution is affected by climate change etc.
- It would be nice to have the map automatically zoom in to the extent of the layers currently being shown (but that's a very minor suggestion).
- The article doesn't mention data accessibility - I imagine that there are various hurdles there with confidentiality and the fact that this is an endangered species. A statement about this may be worthwhile in the text though.
