+++ Makefile
@@ -49,6 +49,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2021_ccs_tracks.rds \
 														data/interim/2021_neaq_aerial_tracks.rds \
 														data/interim/2021_noaa_twin_otter_tracks.rds \
+														data/interim/2021_dfo_cessna_yob_tracks.rds \
 														data/interim/dcs_archived_tracks.rds \
 														data/interim/dcs_live_tracks.rds
 	$(report_error)
@@ -283,6 +284,12 @@ data/interim/2021_noaa_twin_otter_tracks.rds : R/proc_2021_noaa_twin_otter.R dat
 	Rscript $<
 	$(remove_error)
 
+# 2021 DFO Cessna YOB tracks
+data/interim/2021_dfo_cessna_yob_tracks.rds : R/proc_2021_dfo_cessna_yob_tracks.R data/raw/2021_whalemapdata/DFO_Cessna_YOB/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
 # DCS archived tracks
 data/interim/dcs_archived_tracks.rds : R/proc_archived_dcs.R data/raw/dcs/archived/*/*
 	$(report_error)
@@ -339,6 +346,7 @@ data/processed/observations.rds : R/proc_observations.R \
 																	data/interim/2021_ccs_sightings.rds \
 																	data/interim/2021_neaq_aerial_sightings.rds \
 																	data/interim/2021_noaa_twin_otter_sightings.rds \
+																	data/interim/2021_dfo_cessna_yob_sightings.rds \
 																	data/interim/2021_dfo_opportunistic_sightings.rds \
 																	data/interim/sas_sightings.rds \
 																	data/interim/dcs_archived_detections.rds \
@@ -578,6 +586,12 @@ data/interim/2021_neaq_aerial_sightings.rds : R/proc_2021_neaq_aerial.R data/raw
 
 # 2021 noaa twin otter sightings
 data/interim/2021_noaa_twin_otter_sightings.rds : R/proc_2021_noaa_twin_otter.R data/raw/2021_noaa_twin_otter/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# 2021 DFO cessna YOB sightings
+data/interim/2021_dfo_cessna_yob_sightings.rds : R/proc_2021_dfo_cessna_yob_sightings.R data/raw/2021_whalemapdata/DFO_Cessna_YOB/*
 	$(report_error)
 	Rscript $<
 	$(remove_error)
