### 2021-04-15 09bd923
read emails from excel and always add my email
