+++ src/get_boem_data.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# download map data from BOEM
+# process map data from BOEM
 
 # Extract OS name
 unamestr=`uname`
@@ -12,15 +12,6 @@ elif [[ "$unamestr" == 'Darwin' ]]; then
 	DESTDIR=/Users/hansenjohnson/Projects/WhaleMap # local
 	SSHDIR=/Users/hansenjohnson
 fi
-
-# Determine raw data directory
-DATADIR=${DESTDIR}/data/raw/
-
-# Move to data directory
-cd ${DATADIR}
-
-# Sync BOEM development areas Google drive
-rclone sync drive:"BOEM shapefiles" "boem/" --drive-shared-with-me --drive-formats csv --backup-dir backups -v --exclude *.JPG --exclude *.jpg --exclude *.mp4 --exclude *.pdf
 
 # move to project directory
 cd ${DESTDIR}
