### 2021-04-19 b0aa900
silence boem script and remove tidy select
