### 2021-04-19 bfc1d6e
update 2021 CAN measures and DFO area labels
