+++ paper/paper.bib
@@ -21,3 +21,38 @@
 	journal = {Conservation Science and Practice},	
 	language = {en}		
 }
+
+@article{baumgartner:2020,
+	author = {Baumgartner, Mark F. and Bonnell, Julianne and Corkeron, Peter J. and Van Parijs, Sofie M. and Hotchkin, Cara and Hodges, Ben A. and Bort Thornton, Jacqueline and Mensi, Bryan L. and Bruner, Scott M.},
+	title = {Slocum {Gliders} {Provide} {Accurate} {Near} {Real}-{Time} {Estimates} of {Baleen} {Whale} {Presence} {From} {Human}-{Reviewed} {Passive} {Acoustic} {Detection} {Information}},
+	year = {2020},
+	volume = {7},
+	issn = {2296-7745},
+	doi = {10.3389/fmars.2020.00100},
+	journal = {Frontiers in Marine Science},	
+	language = {en}
+}
+
+@article{koubrak:2021,
+	author = {Koubrak, Olga and VanderZwaag, David L. and Worm, Boris},
+	title = {Saving the {North} {Atlantic} right whale in a changing ocean: {Gauging} scientific and law and policy responses},
+	year = {2021},
+	volume = {200},
+	pages = {105109},
+	issn = {0964-5691},
+	doi = {10.1016/j.ocecoaman.2020.105109},
+	journal = {Ocean \& Coastal Management},
+	language = {en}
+}
+
+@article{gervaise:2021,
+	author = {Gervaise, Cédric and Simard, Yvan and Aulanier, Florian and Roy, Nathalie},
+	title = {Optimizing passive acoustic systems for marine mammal detection and localization: {Application} to real-time monitoring north {Atlantic} right whales in {Gulf} of {St}. {Lawrence}},
+	year = {2021},	
+	volume = {178},
+	pages = {107949},
+	issn = {0003-682X},
+	doi = {10.1016/j.apacoust.2021.107949},
+	journal = {Applied Acoustics},
+	language = {en}
+}
