+++ paper/paper.md
@@ -51,7 +51,7 @@ Once the survey data are processed, they are visualized using two different meth
 
 # Conclusions
 
-Since its launch in 2018, `WhaleMap` has been constantly refined and optimized to better serve the overall goal of providing a common source for all near real-time whale survey data in the Northwest Atlantic. It has demonstrably improved conservation outcomes for endangered whales in this region by optimizing research activities, facilitating dynamic risk-mitigation measures, and engaging with the ocean industry and the public. `WhaleMap` has also already been cited in several scientific publications [e.g., @johnson:2020; @kowarski:2020]. It is our hope that `WhaleMap` continues to serve the conservation community in perpetuity, and that this system can be readily adapted to benefit other regions facing similar conservation challenges.
+Since its launch in 2018, `WhaleMap` has been constantly refined and optimized to better serve the overall goal of providing a common source for all near real-time whale survey data in the Northwest Atlantic. It has demonstrably improved conservation outcomes for endangered whales in this region by optimizing research activities, facilitating dynamic risk-mitigation measures, and engaging with the ocean industry and the public. `WhaleMap` has also already been cited in several scientific publications [e.g., @gervaise:2021; @koubrak:2021; @baumgartner:2020; @johnson:2020; @kowarski:2020]. It is our hope that `WhaleMap` continues to serve the conservation community in perpetuity, and that this system can be readily adapted to benefit other regions facing similar conservation challenges.
 
 # Figures
 
