### 2021-04-21 #3094 @hansenjohnson
Hi @pjbouchet, thanks again for your helpful comments! I have taken the following steps to address them:

> Found a typo in the README file, where `src/get_get_remote_data.sh` should be `src/get_remote_data.sh`

[Fixed](https://github.com/hansenjohnson/WhaleMap/commit/51a57348d35138c1bb18c9000c6d561d7cb7ed4c)

> WhaleMap is cited/mentioned in more papers than currently reported in the reference list, though some authors only provide the app URL rather than include an actual citation which could be more easily tracked. You may want to add the following to your list of references:
>
> * [Koubrak et al. 2021](https://www.sciencedirect.com/science/article/pii/S0964569120300193). Saving the North Atlantic right whale in a changing ocean: Gauging scientific and law and policy responses.
> * [Gervaise et al. 2021.](https://www.sciencedirect.com/science/article/pii/S0003682X21000426) Optimizing passive acoustic systems for marine mammal detection and localization: Application to real-time monitoring north Atlantic right whales in Gulf of St. Lawrence.
> * [Baumgartner et al. 2020. ](https://www.frontiersin.org/articles/10.3389/fmars.2020.00100/full) Slocum Gliders Provide Accurate Near Real-Time Estimates of Baleen Whale Presence From Human-Reviewed Passive Acoustic Detection Information.

[Added](https://github.com/hansenjohnson/WhaleMap/commit/4f204e301ed60177b8f67e5ae007b4c09391358e)

> I don't have any specific comments, other than the above. The app is well designed and works well when I tested it online. A couple of thoughts and suggestions for future improvements:
>
> * I wonder about the sustainability of the workflow given the need to have custom scripts written by the curator for each survey team. In my experience, "allowing any team to submit data in nearly any format" (as stated in the article) is a recipe for headaches. What happens if the curator is suddenly unavailable, changes jobs etc.? Are there systems in place to ensure that the running of the app remains unaffected? I would argue that encouraging data standardisation during collection would be worthwhile.

This is an excellent point. Our choice to accept data in nearly any format was one of necessity. As stated in the manuscript: "[this method] is essential for rapid data collection, as survey teams in the field typically lack the time and resources to coerce their data into a common format.". In our experience the only way to get these data quickly was to accept them in any way, shape or form. Fortunately we are converging on a relatively small number of different data formats that remain stable from one year to the next. Where possible we recommend the use of a common format (e.g., simple excel table for sightings and a standard gpx file for tracklines) when first incorporating data from a new group. We are also taking several steps to seek long term support for the system and its curator(s).

> * Are you thinking / have you thought about adding environmental layers to the map (and the possibility to query environmental data at the locations where whales were detected)? This would be interesting and could support spatial modelling efforts to better understand how whale distribution is affected by climate change etc.

Great idea! There are several efforts underway that are interested in pursuing this kind of work (e.g., combining WhaleMap detections and operational ocean model output to forecast whale distribution). We are often approached about adding these and other layers to WhaleMap. In general we are very cautious about adding layers out of concern that they would clutter the display, overwhelm general users, and distract from the core goal of WhaleMap (display Whale survey results). An alternative that we are pursuing is to provide the WhaleMap data feed to other groups (with permission) who could then build a similar viewer and add layers that are specific to their project.

> * It would be nice to have the map automatically zoom in to the extent of the layers currently being shown (but that's a very minor suggestion).

We have pursued this in the past. It is very nice in some instances, but becomes jarring and undesirable when the user wishes to keep the map focused on a specific region while they adjust the data filters. This is a very common use case, so I think we will keep the current behaviour.

> * The article doesn't mention data accessibility - I imagine that there are various hurdles there with confidentiality and the fact that this is an endangered species. A statement about this may be worthwhile in the text though.

This is an important point. The last bullet in the Statement of Need reads: "[Critically, `WhaleMap` does **not**:] Allow access to raw or processed data without approval from the data originator". We feel this adequately covers the issue of data accessibility, but can elaborate if necessary.

> You may want to add a sentence to the paper indicating the lack of equivalent software in existence.

[Added](https://github.com/hansenjohnson/WhaleMap/commit/fd2d70e82577a42a00518f59eda37631ea77925e)

Please let me know if anything needs clarification. Many thanks again for your help, it has certainly improved the manuscript!
