### 2021-04-30 #3094 @mcsiple
Thank you all for your patience-- below are my review notes.

# WhaleMap review
This is a great tool and both the code and ui are simple and clean. An excellent use of `leaflet` for displaying several data layers. I can see how it is already a useful tool for survey teams, but could also be great for teaching about surveys for cetacean abundance.

## Substantial scholarly effort
I know that the app is not intended to provide raw data outputs without approval from the data originators. Technically the app *would* be more likely to be used by scientists like myself if there were some way to either:

1) Direct the user to the pathway for getting raw data if they need it. For example, including contact information where people might dig further (as I'm sure data exploration in this tool will lead to research questions). Or

2) Include, if possible, an option for the app to show *summaries* of sightings data at a coarse enough grain that it would be permissible by the data sources.

These are not deficiencies in the app but would lead to more citations of the tool. I do note that it has been cited by other papers before (@pjbouchet caught the full list) so will leave this detail to the editor to decide.

## Functionality
The app works great and is quite fast. I really only have minor comments on it.

### `server` comments:
I am impressed with how clean and polished both the `server` and `ui` are. Well done; the code is very easy to read and interpret. I had some of the same workflow concerns as @pjbouchet but I think you've addressed them!

### `ui` comments:

- If this tool is used by survey teams to coordinate, do they have access to finer-grain data or raw data if they have a password? That seems like it would be useful for their purposes.
- blue line on upper plot (Acoustic detection events per day) is below zero-- I think this just needs to be moved up to 0. In general it seems safe to fix yaxis start at 0 across all sightings plots.
- "Go!" button is small and hard to see on my monitor -- suggest moving to the sidebar
- might be good to move incomplete data warning (yellow box at bottom right when selected start date is earlier than first complete data year) up to sidebar too (or make larger, or fix, or both).
- If the "Range among years" option shows a multi-year range between the earliest and latest year in the `selectizeInput()` widget, maybe it is more streamlined to make this input a double-ended range slider?
- when I changed the "choose date(s)" option to "Range among years," I got an error when I first tried it (about a week ago; "Error: Evaluation error: wrong sign in 'by' argument") - I haven't been able to get it again but if that happens with other errors I suggest using `safeError()` to build more informative errors into the ui. You can also use `tags$head(tags$style(".shiny-output-error{color: mycolor;}"))` as an argument to `fluidPage()` to change the color of the errors (my experience is that people panic when they get red error messages)
- tiny suggestion: change legend in leaflet to show points for observations instead of plain squares. Or change the grey color for 'buoy' vs 'definite visual'

## Documentation:
@KristinaRiemer, most of my comments in this section are just editor discretion. There are a few listed things on JOSS's checklist that are absent but I think there is good reason those absences.

### General comments

- Installation instructions are NA because the app runs on the server only (I am guessing the authors cannot post it as a R package to run locally because of the connection between the app and the server connection to the data). This is fine with me as long as it's fine with JOSS!

- Usage examples are also NA. Would like to see a short blurb about a real-time usage example for a survey team or an educational example

- Automated tests: I don't actually see any, probably for the 'R package' reason mentioned above. Just want to confirm with the editors that exceptions to this are okay. I think for this particular tool, it is appropriate.

## Software paper

### General quality
My only writing suggestions are to the Statement of Need -- everything else is great. I would like to see a "use case" here.

### Statement of need

- "attempt" seems the wrong term here. Maybe "[baleen whales] have recovered slowly from commercial whaling"

- by 'infrastructure' do you mean shipping, fishing, or both?

- line 2: suggest to just change this to 'shifting environmental baseline' or explicitly say human-driven climate change ('exploitation' in a marine context is almost exlusively used in the context of fisheries)

- It is hard [for me] to identify a broader audience in the scientific community. Anyone besides the survey teams? Vessel captains? The public? Perhaps the intended user groups should be mentioned here.
