+++ server.R
@@ -284,8 +284,8 @@ function(input, output, session){
   # password warnings
   observeEvent(input$go,{
     if(input$password == password){
-      showNotification('Password was correct! Showing unverified and/or test data...',
-                       duration = 7, closeButton = T, type = 'message')
+      showNotification(h4('Password was correct! Showing unverified and/or test data...'),
+                       duration = NULL, closeButton = T, type = 'message')
     }
   })
   
@@ -294,24 +294,24 @@ function(input, output, session){
     
     # track warning
     if(nrow(trk())>npts){
-      showNotification(paste0('Warning! Tracklines have been turned off because 
+      showNotification(h4(paste0('Warning! Tracklines have been turned off because 
                               you have chosen to plot more data than this application 
                               can currently handle (i.e. more than ', as.character(npts), ' points). 
-                              Please select less data to view tracks.'), 
-                       duration = 7, closeButton = T, type = 'warning')
+                              Please select less data to view tracks.')), 
+                       duration = 15, closeButton = T, type = 'error')
     }
     
     # species warning
     if(paste(species(),collapse=',')!='right'){
-      showNotification('Note: WhaleMap focuses on right whales. Other species
-                              information is incomplete.', 
-                       duration = 7, closeButton = T, type = 'warning')
+      showNotification(h4('Note: WhaleMap focuses on right whales. Other species
+                              information is incomplete.'), 
+                       duration = 15, closeButton = T, type = 'error')
     }
     
     # year warning
     if(min(year(dates()))<2017){
-      showNotification('Note: Data before 2017 are incomplete.', 
-                       duration = 7, closeButton = T, type = 'warning')
+      showNotification(h4('Note: Data before 2017 are incomplete.'), 
+                       duration = 15, closeButton = T, type = 'error')
     }
     
   })
