### 2021-05-02 d9b64e3
change buoy color and increase stroke
