### 2021-05-02 hansenjohnson/WhaleMap#63 @hansenjohnson
Feature request: more descriptive error reporting
Making error messages more descriptive could be very helpful for reproducing errors and debugging. There are a number of resources for achieving finer control over Shiny error messages, such as this [article](https://shiny.rstudio.com/articles/sanitize-errors.html). It was recommended I try using `safeError()` specifically. In addition, it is also possible to edit the UI CSS to change the colour of the error messages as follows: `tags$head(tags$style(".shiny-output-error{color: mycolor;}"))`
