### 2021-05-02 #3094 @hansenjohnson
Hi @mcsiple! Thank you very much for your clear and constructive comments. Please see my responses below:

> ## Substantial scholarly effort
> I know that the app is not intended to provide raw data outputs without approval from the data originators. Technically the app _would_ be more likely to be used by scientists like myself if there were some way to either:
>
> 1. Direct the user to the pathway for getting raw data if they need it. For example, including contact information where people might dig further (as I'm sure data exploration in this tool will lead to research questions). Or

We have considered adding contact information for each survey group, but have found that doing so ends up being less efficient for a variety of reasons. For example, it makes it more difficult for a WhaleMap user to compile a comprehensive data request, and adds to the workload of each survey team. We have found it preferable to have all requests for WhaleMap data submitted via email to me (hansen.johnson@dal.ca). That way I can work with my team to review the requests, gather additional information, identify duplication of effort, solicit the appropriate permissions, and pull the requested data.

> 2. Include, if possible, an option for the app to show _summaries_ of sightings data at a coarse enough grain that it would be permissible by the data sources.

This is an excellent suggestion. Many users currently take screenshots of the app for survey summary purposes, but in the future perhaps a more elegant solution would be to add a feature to either take a screenshot automatically (perhaps using ), or even capture the main features of the map, timeseries and summary table in a small summary report. I recently filed an [issue](https://github.com/hansenjohnson/WhaleMap/issues/60) with a very similar feature request. I would not consider this essential to implement prior to publication, but I leave that to the discretion of the editor.

> ## Functionality
>
> ### `ui` comments:
> * If this tool is used by survey teams to coordinate, do they have access to finer-grain data or raw data if they have a password? That seems like it would be useful for their purposes.

Currently, entering a password reveals unverified data such as opportunistic sightings reports or survey platforms that are still in development. It does not provide access to raw data or data at a different resolution. I work quite closely with a number of survey teams that use WhaleMap regularly. In my experience most teams use the measurement tool (at the bottom left of the map), graticules layer, management layers, and (in US waters) NOAA charts, in addition to the data filters, to plan their survey activities. None have suggested to me that access to the raw data would be helpful for survey planning purposes. I also suspect that some teams would not be comfortable allowing their raw data to be accessed by other groups without their knowledge. Any teams requiring raw data for more detailed analyses are welcome to request access.

> * blue line on upper plot (Acoustic detection events per day) is below zero-- I think this just needs to be moved up to 0. In general it seems safe to fix yaxis start at 0 across all sightings plots.

The problem with place this effort line at zero is that it becomes very difficult to see when the y axis range gets large (i.e., when displaying large numbers of observations). The only way I can think to mitigate this issue is to add conditions to change the line position depending on the numbers of observations. I would like to avoid that if possible, as I suspect it would reduce the performance of the application and potentially confuse the viewer. Let me know if I am misunderstanding your suggestion, or if there are obvious alternatives that I may be overlooking.

> * "Go!" button is small and hard to see on my monitor -- suggest moving to the sidebar

I absolutely agree. In the past I had the "Go!" button on the sidebar, but many users became frustrated because they would either forget to click "Go!" after changing the filters, or because they constantly had to scroll up and down between the map, filters, and "Go!" button. It's also important to consider multiple devices / screen sizes for this kind of change, as WhaleMap is often used via smartphone or other small mobile devices. I have made a couple [changes](https://github.com/hansenjohnson/WhaleMap/commit/9b888bc47f8f4cfe25da73688eebe38ed1650a2e) to increase the size of the button, which I already find is a substantial improvement. I hope that resolves your issue as well.

> * might be good to move incomplete data warning (yellow box at bottom right when selected start date is earlier than first complete data year) up to sidebar too (or make larger, or fix, or both).

I have had several conversations recently that make it clear to me that these data warnings need to be made more prominent, so thank you for these suggestions. I have made a few [changes](https://github.com/hansenjohnson/WhaleMap/commit/0a20f215ed7a32d3ff746998eeeaaeb2ed93524b) including increasing the text size, changing the color (to red), and also increasing the length of time the warnings last before they expire.

> * If the "Range among years" option shows a multi-year range between the earliest and latest year in the `selectizeInput()` widget, maybe it is more streamlined to make this input a double-ended range slider?

I prefer the `selectizeInput()` because it allows the comparison of non-sequential years. For example, the `selectizeInput()` allows you to select data specifically from 2015 and 2017 (skipping 2016), rather than 2015-2017 (including 2016). It does make it slightly more cumbersome for users to select many years, but I think this is acceptable, particularly when the "Date range" option makes it very easy to show all data over a large time range.

> * when I changed the "choose date(s)" option to "Range among years," I got an error when I first tried it (about a week ago; "Error: Evaluation error: wrong sign in 'by' argument") - I haven't been able to get it again but if that happens with other errors I suggest using `safeError()` to build more informative errors into the ui. You can also use `tags$head(tags$style(".shiny-output-error{color: mycolor;}"))` as an argument to `fluidPage()` to change the color of the errors (my experience is that people panic when they get red error messages)

I am not sure what precipitated this error. Perhaps some bad data were briefly uploaded in the last few weeks as I have been bringing some new survey platforms online. I am glad it seems to be resolved now, but I will be watchful for it in the future. I like your suggestion to aspire to provide more descriptive error messaging. While these messages probably will not be relevant to most users, they would certainly help me in reproducing and debugging. I think this is a relatively minor change so have filed it as a [feature request](https://github.com/hansenjohnson/WhaleMap/issues/63) and leave the urgency of the implementation up to the editor.

> * tiny suggestion: change legend in leaflet to show points for observations instead of plain squares. Or change the grey color for 'buoy' vs 'definite visual'

Thanks a lot for this. I am slightly colorblind so occasionally struggle with color conflicts like this. I have [changed](https://github.com/hansenjohnson/WhaleMap/commit/d9b64e31cd45663f8dfacae566c6d89d5707cf7f) the default track colour for 'buoy' to better distinguish it from the 'definite visual' detections. Note that the default colours can be adjusted further in the 'Colors' tab of the UI.

> ## Documentation:
>
> ### General comments
> * Usage examples are also NA. Would like to see a short blurb about a real-time usage example for a survey team or an educational example

Excellent idea! I have added a [few sentences](https://github.com/hansenjohnson/WhaleMap/commit/1b88c2e2dc052e92cc41c0297c82d1807c780b01) to the end of the "Statement of Need" that hopefully provide suitable usage examples.

> ## Software paper
> ### General quality
> My only writing suggestions are to the Statement of Need -- everything else is great. I would like to see a "use case" here.

See previous comment.

> ### Statement of need
> * "attempt" seems the wrong term here. Maybe "[baleen whales] have recovered slowly from commercial whaling"

I appreciate that "attempt" is perhaps an anthropomorphism. I used it to convey that none of these species, as far as I am aware, have actually recovered to their pre-whaling abundances. I have made [new edit](https://github.com/hansenjohnson/WhaleMap/commit/1b88c2e2dc052e92cc41c0297c82d1807c780b01) that reads: "[Baleen whale] recovery from commercial whaling is impeded by..."

> * by 'infrastructure' do you mean shipping, fishing, or both?

I used 'infrastructure' as something of a catch-all term that applies to all manner of our physical occupation of the ocean. It is not exclusive to shipping and fishing but also relates to other activities like energy extraction (e.g., wind, petroleum, tide, mining, aquaculture, etc.). I have replaced with "industry", which hopefully captures these ideas more clearly.

> * line 2: suggest to just change this to 'shifting environmental baseline' or explicitly say human-driven climate change ('exploitation' in a marine context is almost exlusively used in the context of fisheries)

Thank you for pointing out that 'exploitation' may have specific connotations and convey an incomplete message. However, I think using only "shifting environmental baseline" or "human-driven climate change" omits the very significant risks posed by industry (e.g., fishing gear entanglement, ship strike) and pollution (e.g., noise, chemicals, plastic). I have [edited](https://github.com/hansenjohnson/WhaleMap/commit/1b88c2e2dc052e92cc41c0297c82d1807c780b01) so the sentence now reads: "[Baleen whale] recovery from commercial whaling is impeded by anthropogenic risks from ocean industry, pollution, and climate change".

> * It is hard [for me] to identify a broader audience in the scientific community. Anyone besides the survey teams? Vessel captains? The public? Perhaps the intended user groups should be mentioned here.

Hopefully this has been addressed in the new usage examples included at the end of the "Statement of Need".

Thank you again for your thoughtful review. It has greatly improved the manuscript and application!
