### 2021-05-03 4247ae5
fix minor typos/redundancies
