+++ paper/paper.md
@@ -41,7 +41,7 @@ The `WhaleMap` system workflow can be separated into data processing and visuali
 
 ## Data processing
 
-Survey teams provide `WhaleMap` access to a remote repository of their choice (e.g., Google Drive, Dropbox, etc.) where they upload their survey data. The `WhaleMap` curator writes a custom script to extract the detection and effort data from each survey team and convert it to a common `WhaleMap` format. This method eases the burden on the survey teams by allowing any team to submit data in nearly any format, provided the format is consistent and well-documented. This is essential for rapid data collection, as survey teams in the field typically lack the time and resources to coerce their data into a common format.
+Survey teams provide `WhaleMap` access to a remote repository of their choice (e.g., Google Drive, Dropbox) where they upload their survey data. The `WhaleMap` curator writes a custom script to extract the detection and effort data from each survey team and convert it to a common `WhaleMap` format. This method eases the burden on the survey teams by allowing any team to submit data in nearly any format, provided the format is consistent and well-documented. This is essential for rapid data collection, as survey teams in the field typically lack the time and resources to reformat their data.
 
 A scheduled job regularly clones the data from the remote repositories onto the `WhaleMap` server and uses a makefile to dynamically and efficiently process the data from each platform and coerce it into a common format. Formatting errors in a remote data repository are automatically flagged and the contributor is notified. This ensures that any changes to the raw survey data quickly propagate through the entire system, which allows survey teams to retain complete control of their data and perform quality control as needed. It also guarantees that the `WhaleMap` always contains the latest available information.
 
