+++ Makefile
@@ -50,6 +50,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2021_neaq_aerial_tracks.rds \
 														data/interim/2021_noaa_twin_otter_tracks.rds \
 														data/interim/2021_dfo_cessna_yob_tracks.rds \
+														data/interim/2021_tc_dash8_tracks.rds \
 														data/interim/dcs_archived_tracks.rds \
 														data/interim/dcs_live_tracks.rds
 	$(report_error)
@@ -286,6 +287,12 @@ data/interim/2021_noaa_twin_otter_tracks.rds : R/proc_2021_noaa_twin_otter.R dat
 
 # 2021 DFO Cessna YOB tracks
 data/interim/2021_dfo_cessna_yob_tracks.rds : R/proc_2021_dfo_cessna_yob_tracks.R data/raw/2021_whalemapdata/DFO_Cessna_YOB/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# 2021 TC dash8 tracks
+data/interim/2021_tc_dash8_tracks.rds : R/proc_2021_tc_dash8_tracks.R data/raw/2021_whalemapdata/TC_Dash8/*
 	$(report_error)
 	Rscript $<
 	$(remove_error)
