+++ Makefile
@@ -51,6 +51,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2021_noaa_twin_otter_tracks.rds \
 														data/interim/2021_dfo_cessna_yob_tracks.rds \
 														data/interim/2021_tc_dash8_tracks.rds \
+														data/interim/2021_dfo_twin_otter_tracks.rds \
 														data/interim/dcs_archived_tracks.rds \
 														data/interim/dcs_live_tracks.rds
 	$(report_error)
@@ -297,6 +298,12 @@ data/interim/2021_tc_dash8_tracks.rds : R/proc_2021_tc_dash8_tracks.R data/raw/2
 	Rscript $<
 	$(remove_error)
 
+# 2021 DFO twin otter tracks
+data/interim/2021_dfo_twin_otter_tracks.rds : R/proc_2021_dfo_twin_otter_tracks.R data/raw/2021_whalemapdata/DFO_twin_otter/*/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
 # DCS archived tracks
 data/interim/dcs_archived_tracks.rds : R/proc_archived_dcs.R data/raw/dcs/archived/*/*
 	$(report_error)
@@ -355,6 +362,7 @@ data/processed/observations.rds : R/proc_observations.R \
 																	data/interim/2021_noaa_twin_otter_sightings.rds \
 																	data/interim/2021_dfo_cessna_yob_sightings.rds \
 																	data/interim/2021_tc_dash8_sightings.rds \
+																	data/interim/2021_dfo_twin_otter_sightings.rds \
 																	data/interim/2021_dfo_opportunistic_sightings.rds \
 																	data/interim/sas_sightings.rds \
 																	data/interim/dcs_archived_detections.rds \
@@ -606,6 +614,12 @@ data/interim/2021_dfo_cessna_yob_sightings.rds : R/proc_2021_dfo_cessna_yob_sigh
 
 # 2021 tc dash8 sightings
 data/interim/2021_tc_dash8_sightings.rds : R/proc_2021_tc_dash8_sightings.R data/raw/2021_whalemapdata/TC_Dash8/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# 2021 DFO twin otter sightings
+data/interim/2021_dfo_twin_otter_sightings.rds : R/proc_2021_dfo_twin_otter_sightings.R data/raw/2021_whalemapdata/DFO_twin_otter/*/*
 	$(report_error)
 	Rscript $<
 	$(remove_error)
