+++ Makefile
@@ -52,6 +52,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2021_dfo_cessna_yob_tracks.rds \
 														data/interim/2021_tc_dash8_tracks.rds \
 														data/interim/2021_dfo_twin_otter_tracks.rds \
+														data/interim/2021_dfo_cessna_zwf_tracks.rds \
 														data/interim/dcs_archived_tracks.rds \
 														data/interim/dcs_live_tracks.rds
 	$(report_error)
@@ -304,6 +305,12 @@ data/interim/2021_dfo_twin_otter_tracks.rds : R/proc_2021_dfo_twin_otter_tracks.
 	Rscript $<
 	$(remove_error)
 
+# 2021 DFO Cessna ZWF tracks
+data/interim/2021_dfo_cessna_zwf_tracks.rds : R/proc_2021_dfo_cessna_zwf_tracks.R data/raw/2021_whalemapdata/DFO_Cessna_ZWF/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
 # DCS archived tracks
 data/interim/dcs_archived_tracks.rds : R/proc_archived_dcs.R data/raw/dcs/archived/*/*
 	$(report_error)
@@ -364,6 +371,7 @@ data/processed/observations.rds : R/proc_observations.R \
 																	data/interim/2021_tc_dash8_sightings.rds \
 																	data/interim/2021_dfo_twin_otter_sightings.rds \
 																	data/interim/2021_dfo_opportunistic_sightings.rds \
+																	data/interim/2021_dfo_cessna_zwf_sightings.rds \
 																	data/interim/sas_sightings.rds \
 																	data/interim/dcs_archived_detections.rds \
 																	data/interim/dcs_live_detections.rds
@@ -620,6 +628,12 @@ data/interim/2021_tc_dash8_sightings.rds : R/proc_2021_tc_dash8_sightings.R data
 
 # 2021 DFO twin otter sightings
 data/interim/2021_dfo_twin_otter_sightings.rds : R/proc_2021_dfo_twin_otter_sightings.R data/raw/2021_whalemapdata/DFO_twin_otter/*/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# 2021 DFO cessna ZWF sightings
+data/interim/2021_dfo_cessna_zwf_sightings.rds : R/proc_2021_dfo_cessna_zwf_sightings.R data/raw/2021_whalemapdata/DFO_Cessna_ZWF/*
 	$(report_error)
 	Rscript $<
 	$(remove_error)
