+++ R/proc_gis.R
@@ -39,10 +39,19 @@ dfo_zone@data$ID = c(
 critical_habitat_zone = readOGR('data/raw/2021_whalemapdata/GIS_data/critical_habitat_areas/') %>%
   spTransform(ref)
 
+# read DFO lines
+dfo_10f = readOGR('data/raw/2021_whalemapdata/GIS_data/2021_DFO_10 and 20 fathom lines/', layer = 'Fathom_10') %>%
+  spTransform(ref)
+dfo_20f = readOGR('data/raw/2021_whalemapdata/GIS_data/2021_DFO_10 and 20 fathom lines/', layer = 'Fathom_20') %>%
+  spTransform(ref)
+
 # read management grid
 full_grid = readOGR('data/raw/2021_whalemapdata/GIS_data/Full_ATL_grids-2021/') %>%
   spTransform(ref)
 full_grid@data$Grid_Index = full_grid@data$GridName
+full_grid@data$GridName = NULL
+
+# crop to >40 lat
 
 # read US lobster zones
 us_lobster0 = readOGR('data/raw/gis/Lobster_Management_Areas/') %>%
@@ -59,6 +68,8 @@ us_lobster = SpatialPolygonsDataFrame(us_lobster, data=us_lobster0@data)
 #   addLayersControl(
 #     overlayGroups = c('tc_zone',
 #                       'dfo_zone',
+#                       'dfo_10f',
+#                       'dfo_20f',
 #                       'tc_ra',
 #                       'critical_habitat_zone',
 #                       'us_lobster')
@@ -67,6 +78,10 @@ us_lobster = SpatialPolygonsDataFrame(us_lobster, data=us_lobster0@data)
 #                group = 'tc_zone',weight = 2) %>%
 #   addPolygons(data = dfo_zone, color = 'orange', popup = ~paste0(ID),
 #               group = 'dfo_zone',weight = 2) %>%
+#   addPolylines(data = dfo_10f, color = 'black',
+#               group = 'dfo_10f',weight = 2) %>%
+#   addPolylines(data = dfo_20f, color = 'grey',
+#               group = 'dfo_20f',weight = 2) %>%
 #   addPolygons(data = tc_ra, color = 'grey',
 #               group = 'tc_ra',weight = 2) %>%
 #   addPolygons(data = us_lobster, color = 'red', popup = ~paste0(COMMNAME),
@@ -77,6 +92,8 @@ us_lobster = SpatialPolygonsDataFrame(us_lobster, data=us_lobster0@data)
 # save
 save(tc_zone,
      dfo_zone,
+     dfo_10f,
+     dfo_20f,
      critical_habitat_zone,
      full_grid,
      tc_ra,
