+++ R/proc_2021_tc_dash7_sightings.R
@@ -1,0 +1,127 @@
+## proc_2021_tc_dash7_sightings ##
+# Process sightings data from TC Dash-7 survey plane
+
+# user input --------------------------------------------------------------
+
+# data directory
+data_dir = 'data/raw/2021_whalemapdata/TC_Dash7/'
+
+# output file name
+ofile = 'data/interim/2021_tc_dash7_sightings.rds'
+
+# setup -------------------------------------------------------------------
+
+# libraries
+suppressPackageStartupMessages(library(readxl))
+
+# functions
+source('R/functions.R')
+
+# list files to process
+flist = list.files(data_dir, pattern = '(\\d{8})_Dash7_sightings.xls.$', full.names = T, recursive = T)
+
+# only proceed if sightings files exist
+if(length(flist)!=0){
+  
+  # list to hold loop output
+  SIG = vector('list', length = length(flist))
+  
+  # read and format data ----------------------------------------------------
+  
+  # read files
+  for(i in seq_along(flist)){
+    
+    # skip empty files
+    if (file.size(flist[i]) == 0) next
+    
+    # read in data
+    tmp = as.data.frame(read_excel(flist[i], sheet = 1, col_names = TRUE))
+    
+    # check for position data
+    if(!'Pos_lat' %in% names(tmp)){
+      message('Skipping file without position data: ', flist[i])
+      next
+    }
+    
+    # check for sightings data
+    if(length(tmp$Pos_lat)==1 & is.na(tmp$Pos_lat[1])){
+      message('Skipping file without sightings data: ', flist[i])
+      next
+    }
+    
+    # select columns of interest
+    tmp = data.frame(date = tmp$Date_UTC,
+                     lat = tmp$Pos_lat,
+                     lon = tmp$Pos_long,
+                     time = tmp$Time_UTC,
+                     species = tmp$Sp_code,
+                     number = tmp$Nb_total
+    )
+    
+    
+    # remove columns without data
+    tmp = tmp[!is.na(tmp$species)&
+                !is.na(tmp$lat)&
+                !is.na(tmp$lon)&
+                !is.na(tmp$time),]
+    
+    # fix date
+    tmp$date = as.Date(tmp$date[1]) # only use first date
+    
+    # skip if nothing seen
+    if (nrow(tmp) == 0) next
+    
+    # fix time
+    tmp$time = as.POSIXct(tmp$time, tz = 'UTC')
+    tmp$time = as.POSIXct(paste0(tmp$date, ' ', hour(tmp$time), ':', minute(tmp$time), ':', second(tmp$time)), 
+                          tz = 'UTC', usetz = T)
+    
+    # convert lat lon data type
+    tmp$lat = ddm2dd_col(as.character(tmp$lat))
+    tmp$lon = -ddm2dd_col(as.character(tmp$lon))
+    
+    # add species identifiers
+    tmp$species = toupper(tmp$species)
+    tmp$species[tmp$species == 'EG'] = 'right'
+    tmp$species[tmp$species == 'MN'] = 'humpback'
+    tmp$species[tmp$species == 'BB'] = 'sei'
+    tmp$species[tmp$species == 'BP'] = 'fin'
+    tmp$species[tmp$species == 'FS'] = 'fin/sei'
+    tmp$species[tmp$species == 'BA'] = 'minke'
+    tmp$species[tmp$species == 'BM'] = 'blue'
+    tmp$species[tmp$species == 'LWNR'] = 'unknown rorqual'
+    tmp$species[tmp$species == 'UnWh'] = 'unknown large whale'
+    tmp$species[tmp$species == 'UC'] = 'unknown cetacean'
+    
+    # add metadata
[…]
