### 2021-05-19 3ea7b63
explicitly specify tc sighting file name convention
