+++ R/proc_2021_tc_dash8_sightings.R
@@ -23,7 +23,7 @@ suppressPackageStartupMessages(library(stringr))
 source('R/functions.R')
 
 # list files to process
-flist = list.files(data_dir, pattern = '(\\d{8})_Dash8_sightings.xls.$', full.names = T, recursive = T)
+flist = list.files(data_dir, pattern = '^(\\d{8})_Dash8_sightings.xls.$', full.names = T, recursive = T)
 
 # only proceed if sightings files exist
 if(length(flist)!=0){
