+++ R/proc_2021_tc_dash7_sightings.R
@@ -18,7 +18,7 @@ suppressPackageStartupMessages(library(readxl))
 source('R/functions.R')
 
 # list files to process
-flist = list.files(data_dir, pattern = '(\\d{8})_Dash7_sightings.xls.$', full.names = T, recursive = T)
+flist = list.files(data_dir, pattern = '^(\\d{8})_Dash7_sightings.xls.$', full.names = T, recursive = T)
 
 # only proceed if sightings files exist
 if(length(flist)!=0){
