### 2021-05-21 44946bd
extend date range back to 2010
