### 2021-05-21 3ace199
temporary fix for adding 'source' field
