### 2021-05-22 f33ebce
external count for faster barplotting
