### 2021-05-22 1f28b45
catch no data error in text summary
