+++ Makefile
@@ -34,6 +34,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2019_mics_tracks.rds \
 														data/interim/2019_neaq_tracks.rds \
 														data/interim/2019_dfo_cp_tracks.rds \
+														data/interim/2019_dfo_coriolis_tracks.rds \
 														data/interim/2019_dfo_twin_otter_tracks.rds \
 														data/interim/2020_noaa_twin_otter_tracks.rds \
 														data/interim/2020_dfo_cessna_tracks.rds \
@@ -198,6 +199,12 @@ data/interim/2019_dfo_twin_otter_tracks.rds : R/proc_2019_dfo_twin_otter_tracks.
 	Rscript $<
 	$(remove_error)
 
+# 2019 DFO Coriolis
+data/interim/2019_dfo_coriolis_tracks.rds : R/proc_2019_dfo_coriolis_tracks.R data/raw/2019_whalemapdata/DFO_Coriolis/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
 # 2020 NOAA twin otter tracks
 data/interim/2020_noaa_twin_otter_tracks.rds : R/proc_2020_noaa_twin_otter.R data/raw/2020_noaa_twin_otter/*
 	$(report_error)
@@ -357,6 +364,7 @@ data/processed/observations.rds : R/proc_observations.R \
 																	data/interim/2019_tc_dash7_sightings.rds \
 																	data/interim/2019_mics_sightings.rds \
 																	data/interim/2019_neaq_sightings.rds \
+																	data/interim/2019_dfo_coriolis_sightings.rds \
 																	data/interim/2020_noaa_twin_otter_sightings.rds \
 																	data/interim/2020_dfo_cessna_sightings.rds \
 																	data/interim/2020_dfo_cessna2_sightings.rds \
@@ -512,6 +520,12 @@ data/interim/2019_mics_sightings.rds : R/proc_2019_mics_sightings.R data/raw/201
 
 # 2019 NEAQ sightings
 data/interim/2019_neaq_sightings.rds : R/proc_2019_neaq.R data/raw/2019_neaq/*.csv
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# 2019 DFO Corliolis sightings
+data/interim/2019_dfo_coriolis_sightings.rds : R/proc_2019_dfo_coriolis_sightings.R data/raw/2019_whalemapdata/DFO_Coriolis/*
 	$(report_error)
 	Rscript $<
 	$(remove_error)
