+++ Makefile
@@ -55,6 +55,7 @@ data/processed/tracks.rds : R/proc_tracks.R \
 														data/interim/2021_dfo_twin_otter_tracks.rds \
 														data/interim/2021_dfo_cessna_zwf_tracks.rds \
 														data/interim/2021_tc_dash7_tracks.rds \
+														data/interim/2021_dfo_viking_tracks.rds \
 														data/interim/dcs_archived_tracks.rds \
 														data/interim/dcs_live_tracks.rds
 	$(report_error)
@@ -235,12 +236,6 @@ data/interim/2020_dfo_twin_otter_tracks.rds : R/proc_2020_dfo_twin_otter_tracks.
 	Rscript $<
 	$(remove_error)
 
-# DFO viking tracks
-#data/interim/2020_dfo_viking_tracks.rds : R/proc_2020_dfo_viking.R
-#	$(report_error)
-#	Rscript $<
-#	$(remove_error)
-
 # 2020 Hawkins tracks
 data/interim/2020_niha_tracks.rds : R/proc_2020_niha_tracks.R data/raw/2020_niha/Tracks/*
 	$(report_error)
@@ -325,6 +320,12 @@ data/interim/2021_tc_dash7_tracks.rds : R/proc_2021_tc_dash7_tracks.R data/raw/2
 	Rscript $<
 	$(remove_error)
 
+# DFO viking tracks
+data/interim/2021_dfo_viking_tracks.rds : R/proc_2021_dfo_viking.R
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
 # DCS archived tracks
 data/interim/dcs_archived_tracks.rds : R/proc_archived_dcs.R data/raw/dcs/archived/*/*
 	$(report_error)
@@ -389,9 +390,9 @@ data/processed/observations.rds : R/proc_observations.R \
 																	data/interim/2021_dfo_cessna_zwf_sightings.rds \
 																	data/interim/2021_dfo_cp_sightings.rds \
 																	data/interim/2021_tc_dash7_sightings.rds \
-																	data/interim/2021_dfo_viking_detections.rds \
 																	data/interim/sas_sightings.rds \
 																	data/interim/narwc_sightings.rds \
+																	data/interim/2021_dfo_viking_detections.rds \
 																	data/interim/dcs_archived_detections.rds \
 																	data/interim/dcs_live_detections.rds
 
@@ -573,12 +574,6 @@ data/interim/2020_dfo_twin_otter_sightings.rds : R/proc_2020_dfo_twin_otter_sigh
 	Rscript $<
 	$(remove_error)
 
-# DFO viking detections
-data/interim/2021_dfo_viking_detections.rds : R/proc_2021_dfo_viking.R
-	$(report_error)
-	Rscript $<
-	$(remove_error)
-
 # 2020 coast guard opportunistic sightings
 data/interim/2020_ccgs_sightings.rds : R/proc_2020_ccgs_sightings.R data/raw/2020_whalemapdata/2020-CCGS/*
 	$(report_error)
@@ -689,6 +684,12 @@ data/interim/sas_sightings.rds : R/proc_sas.R
 
 # NARWC sightings
 data/interim/narwc_sightings.rds : R/proc_narwc_sightings.R data/raw/narwc/*
+	$(report_error)
+	Rscript $<
+	$(remove_error)
+
+# DFO viking detections
+data/interim/2021_dfo_viking_detections.rds : R/proc_2021_dfo_viking.R
 	$(report_error)
 	Rscript $<
 	$(remove_error)
