+++ src/get_live_dcs.sh
@@ -18,11 +18,10 @@ declare -A URL
 
 # assign paths to detection data for each deployment
 URL=(
+  [2021-05-25_buoy_nybse]=http://dcs.whoi.edu/nybse0521/nybse0521_nybse_html/ptracks/manual_analysis.csv
+  [2021-05-25_buoy_nybnw]=http://dcs.whoi.edu/nybnw0521/nybnw0521_nybnw_html/ptracks/manual_analysis.csv
 	[2021-03-25_slocum_we15]=http://dcs.whoi.edu/cox0321/cox0321_we15_html/ptracks/manual_analysis.csv
 	[2021-03-17_slocum_we16]=http://dcs.whoi.edu/sbnms0321/sbnms0321_we16_html/ptracks/manual_analysis.csv
-	[2020-12-20_slocum_we03]=http://dcs.whoi.edu/gom1220/gom1220_we03_html/ptracks/manual_analysis.csv
-	[2020-12-03_buoy_ncch]=http://dcs.whoi.edu/ncch1220/ncch1220_ncch_html/ptracks/manual_analysis.csv
-	[2020-01-15_buoy_nybnw]=http://dcs.whoi.edu/nybnw0120/nybnw0120_buoy_html/ptracks/manual_analysis.csv
 )
 
 # download data
