### 2021-06-07 #3094 @mcsiple
I think all my comments are all addressed! The only lingering one is that dotted line at zero sightings on the y-axis-- I suggest just bringing that line to the front and making it a contrasting color if it's hard to see, or making the bars in that plot lighter so it's easier to see them at the same time. From a data viz perspective, I think it is confusing to put a possible number of sightings that could be < 0.

I like the other `ui` changes you made.

Everything looks great! Congratulations @hansenjohnson et al. for making this great tool available.  I accept it!
