### 2021-06-08 #3094 @KristinaRiemer
@hansenjohnson the only change to make to the paper would be to change "html" to "HTML" in line 53. Otherwise the paper and its references look good!
