+++ paper/paper.md
@@ -47,7 +47,7 @@ A scheduled job regularly clones the data from the remote repositories onto the 
 
 ## Visualization
 
-Once the survey data are processed, they are visualized using two different methods. The first is the construction of self-contained html summary maps containing sufficient information to satisfy most casual viewers (typically the last 14-days of survey results). These can be conveniently embedded in various webpages (e.g., [whalemap.org](https://whalemap.org)) and browsed without requiring server-side processing. These maps are dynamically regenerated as the final step in the data processing workflow, so they always contain the latest available information. The second visualization method is an interactive online application ([whalemap.org/WhaleMap](https://whalemap.org/WhaleMap/)). This provides users with numerous tools with which to filter the latest processed data. The selected data are displayed in several formats including an interactive map, interactive timeseries plot, and table of summary statistics. 
+Once the survey data are processed, they are visualized using two different methods. The first is the construction of self-contained HTML summary maps containing sufficient information to satisfy most casual viewers (typically the last 14-days of survey results). These can be conveniently embedded in various webpages (e.g., [whalemap.org](https://whalemap.org)) and browsed without requiring server-side processing. These maps are dynamically regenerated as the final step in the data processing workflow, so they always contain the latest available information. The second visualization method is an interactive online application ([whalemap.org/WhaleMap](https://whalemap.org/WhaleMap/)). This provides users with numerous tools with which to filter the latest processed data. The selected data are displayed in several formats including an interactive map, interactive timeseries plot, and table of summary statistics. 
 
 # Conclusions
 
