### 2021-06-08 #3094 @hansenjohnson
Good catch @KristinaRiemer! I have made that change [here](https://github.com/hansenjohnson/WhaleMap/commit/e388cd161ee5f71cfec3adf3052ca2d88d857018). What are the next steps?
