### 2021-06-08 #3094 @KristinaRiemer
Next steps:

1. Create a tagged release of the GitHub repo (instructions [here](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release)) and post the version number here
2. Generate a DOI by archiving the code on Zenodo, figshare, an institutional repository, etc. and provide that DOI. Make sure the archive has the same title as the paper title, and the same authors.
