### 2021-06-08 #3094 @hansenjohnson
1. The tagged release is version [1.0.0](https://github.com/hansenjohnson/WhaleMap/releases/tag/1.0.0)
2. The DOI is 10.5281/zenodo.4913270

Let me know if there are any issues!
