### 2021-06-08 #3094 @KristinaRiemer
For the Zenodo archive, could you add the other authors as authors there? And the title should match the submission's title of "WhaleMap: a tool to collate and display whale survey1results in near real-time". Could you let me know when those are updated?
