### 2021-06-08 #3094 @kthyng
The paper looks good though a question: what are the stars at the bottom of the first page, and I don't think you need a section "Figures".
