### 2021-06-09 70cb5e9
remove figures section
