### 2021-06-09 #3094 @hansenjohnson
Thanks @kthyng! I have removed the "Figures" sections as well as the stars at the bottom of the page. Let me know if anything else is necessary!
