### 2021-06-09 #3094 @hansenjohnson
Fantastic! Thank you so much again to all of you for your time, feedback, and guidance through this process. Your efforts have greatly improved the manuscript and application. Submitting to JOSS was a wonderful experience!
